+++ src/main/java/com/example/BookingApp/reservation/model/Reservation.java
@@ -23,7 +23,7 @@ public class Reservation {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @ReservationNumber
-    private Integer reservationNumber;
+    private Long reservationNumber;
     @NotBlank
     private Long userId;
     @NotBlank
